+++ Tests/MapboxNavigationTests/NavigationViewControllerTests.swift
@@ -65,7 +65,7 @@ class NavigationViewControllerTests: XCTestCase {
         ResourceOptionsManager.default.resourceOptions.accessToken = .mockedAccessToken
         DirectionsCredentials.injectSharedToken(.mockedAccessToken)
         initialRoute = Fixture.route(from: jsonFileName, options: routeOptions)
-        newRoute = Fixture.route(from: jsonFileName, options: routeOptions)
+        newRoute = Fixture.route(from: "route-with-banner-instructions", options: routeOptions)
         dependencies = {
             UNUserNotificationCenter.replaceWithMock()
 
@@ -280,12 +280,14 @@ class NavigationViewControllerTests: XCTestCase {
             navigationViewController.navigationMapView?.pointAnnotationManager != nil
         }
         waitForExpectations(timeout: 5, handler: nil)
-        navigationViewController.indexedRoute = (initialRoute, 0)
+        navigationViewController.navigationService.router.updateRoute(with: (initialRoute, 0), routeOptions: nil)
 
         expectation(description: "Annotations loaded") {
             !navigationViewController.navigationMapView!.pointAnnotationManager!.annotations.isEmpty
         }
         waitForExpectations(timeout: 5, handler: nil)
+        XCTAssertEqual(navigationViewController.router.routeProgress.route.routeIdentifier, initialRoute.routeIdentifier)
+
 
         let annotations = navigationViewController.navigationMapView!.pointAnnotationManager!.annotations
 
@@ -299,7 +301,7 @@ class NavigationViewControllerTests: XCTestCase {
                   "Destination annotation does not exist on map")
         
         // Set the second route.
-        navigationViewController.indexedRoute = (newRoute, 0)
+        navigationViewController.navigationService.router.updateRoute(with: (newRoute, 0), routeOptions: nil)
         
         let newAnnotations = navigationViewController.navigationMapView!.pointAnnotationManager!.annotations
         
@@ -312,6 +314,8 @@ class NavigationViewControllerTests: XCTestCase {
                     .compactMap { $0.feature.geometry.value as? Turf.Point }
                     .contains { $0.coordinates.distance(to: secondDestination) < 1 },
                   "New destination annotation does not exist on map")
+        XCTAssertEqual(navigationViewController.router.routeProgress.route.routeIdentifier,
+                       newRoute.routeIdentifier)
     }
     
     func testPuck3DLayerPosition() {